move more code into the library for testability
