+++ mobilecoind-dev-faucet/src/data_types.rs
@@ -7,33 +7,53 @@ use mc_util_serial::JsonU64;
 use serde_derive::{Deserialize, Serialize};
 use std::collections::HashMap;
 
+/// A request to the faucet to fund an address
 #[derive(Debug, Default, Deserialize, Serialize)]
 pub struct JsonFaucetRequest {
+    /// The address to fund
     pub b58_address: String,
+    /// The token id to fund. Assumed 0 if omitted.
     pub token_id: Option<JsonU64>,
 }
 
+/// A response describing the status of the faucet server
 #[derive(Debug, Default, Deserialize, Serialize)]
 pub struct JsonFaucetStatus {
+    /// Whether the status request was successful
     pub success: bool,
+    /// The error message in case of failure
     #[serde(skip_serializing_if = "Option::is_none")]
     pub err_str: Option<String>,
+    /// The b58 address of the faucet. This address can be paid to replenish the
+    /// faucet.
     #[serde(skip_serializing_if = "String::is_empty")]
     pub b58_address: String,
+    /// The map of token id -> payout amount for that token id. (The recipient
+    /// gets a little less because of fees.)
     #[serde(skip_serializing_if = "HashMap::is_empty")]
     pub faucet_payout_amounts: HashMap<JsonU64, JsonU64>,
+    /// The current balances of the faucet.
     #[serde(skip_serializing_if = "HashMap::is_empty")]
     pub balances: HashMap<JsonU64, JsonU64>,
+    /// The current depths of the queue of utxos for each token id. If these
+    /// queues run out then the faucet needs some more time to rebuild them.
     #[serde(skip_serializing_if = "HashMap::is_empty")]
     pub queue_depths: HashMap<JsonU64, JsonU64>,
 }
 
+/// A Tx receipt that the reciepient of a payment can use (with mobilecoind)
+/// to track the payment. This is returned with faucet payment responses.
 #[derive(Debug, Default, Deserialize, Serialize)]
 pub struct JsonReceiverTxReceipt {
+    /// The recipient of the payment
     pub recipient: JsonPublicAddress,
+    /// The hex-encoded bytes of the tx out public key
     pub tx_public_key: String,
+    /// The hex-encoded bytes of the tx out hash
     pub tx_out_hash: String,
+    /// The tombstone block of the submitted transaction
     pub tombstone: u64,
+    /// The hex-encoded bytes of the confirmation number
     pub confirmation_number: String,
 }
 
@@ -49,6 +69,7 @@ impl From<&mc_mobilecoind_api::ReceiverTxReceipt> for JsonReceiverTxReceipt {
     }
 }
 
+/// A Json encoded public address structure
 #[derive(Clone, Debug, Default, Deserialize, Serialize)]
 pub struct JsonPublicAddress {
     /// Hex encoded compressed ristretto bytes
@@ -89,9 +110,13 @@ impl From<&PublicAddress> for JsonPublicAddress {
 /// the faucet user cannot make use of the sender tx receipt.
 #[derive(Debug, Default, Deserialize, Serialize)]
 pub struct JsonSubmitTxResponse {
+    /// Whether the payment was submitted successfully
     pub success: bool,
+    /// An error message if the payment could not be submitted successfully
     #[serde(skip_serializing_if = "Option::is_none")]
     pub err_str: Option<String>,
+    /// A receipt for each TxOut that was sent (just one, if submitted
+    /// successfully)
     #[serde(skip_serializing_if = "Vec::is_empty")]
     pub receiver_tx_receipt_list: Vec<JsonReceiverTxReceipt>,
 }
